_legacy_v1/infra/functions/webhook_notifier/requirements.txt
functions-framework==3.8.1
google-cloud-secret-manager==2.21.1
requests==2.32.3
pytz==2024.2